Delete flat vis + move HierarchyControls to its own module.
- Flat vis is irrelevant now that we have UI control over the data
  hierarchy
